+++ src/main/java/com/example/ssg_tab/domain/user/service/UserQueryService.java
@@ -1,0 +1,9 @@
+package com.example.ssg_tab.domain.user.service;
+
+import com.example.ssg_tab.domain.user.dto.response.UserResponse;
+
+public interface UserQueryService {
+
+    UserResponse.UserInfoResponse getUserInfo(Long userId);
+
+}
