+++ src/main/java/com/example/ssg_tab/domain/user/service/UserQueryServiceImpl.java
@@ -1,0 +1,29 @@
+package com.example.ssg_tab.domain.user.service;
+
+import com.example.ssg_tab.domain.user.converter.UserConverter;
+import com.example.ssg_tab.domain.user.dto.response.UserResponse;
+import com.example.ssg_tab.domain.user.entity.User;
+import com.example.ssg_tab.domain.user.repository.UserRepository;
+import com.example.ssg_tab.global.apiPayload.status.ErrorStatus;
+import com.example.ssg_tab.global.exception.GeneralException;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@RequiredArgsConstructor
+public class UserQueryServiceImpl implements UserQueryService {
+
+    private final UserRepository userRepository;
+
+    @Override
+    @Transactional(readOnly = true)
+    public UserResponse.UserInfoResponse getUserInfo(Long userId) {
+
+        User user = userRepository.findById(userId)
+                .orElseThrow(() -> new GeneralException(ErrorStatus.MEMBER_NOT_FOUND));
+
+        return UserConverter.toUserInfoResponse(user);
+    }
+
+}
