+++ src/main/java/com/example/ssg_tab/global/auth/service/AuthService.java
@@ -1,17 +1,13 @@
 package com.example.ssg_tab.global.auth.service;
 
-import com.example.ssg_tab.domain.category.repository.CategoryRepository;
-import com.example.ssg_tab.domain.user.repository.UserCategoryRepository;
-import com.example.ssg_tab.domain.user.service.UserService;
+import com.example.ssg_tab.domain.user.service.UserCreateService;
 import com.example.ssg_tab.global.apiPayload.status.ErrorStatus;
 import com.example.ssg_tab.global.auth.KakaoClient;
 import com.example.ssg_tab.global.exception.GeneralException;
 import com.example.ssg_tab.global.jwt.JwtTokenService;
-import com.example.ssg_tab.global.auth.info.KakaoUserInfo;
 import com.example.ssg_tab.global.auth.converter.AuthConverter;
 import com.example.ssg_tab.global.auth.dto.AuthRequest;
 import com.example.ssg_tab.global.auth.dto.AuthResponse;
-import com.example.ssg_tab.global.auth.info.KakaoTokenInfo;
 import com.example.ssg_tab.domain.user.entity.User;
 import com.example.ssg_tab.domain.user.repository.UserRepository;
 import com.example.ssg_tab.global.jwt.RefreshTokenStore;
@@ -20,8 +16,6 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -34,7 +28,7 @@ public class AuthService {
     private final KakaoClient kakaoClient;
 
     private final UserRepository userRepository;
-    private final UserService userService;
+    private final UserCreateService userCreateService;
 
     @Value("${jwt.refresh-token-exp-days}") private long refreshExpDays;
 
@@ -42,10 +36,10 @@ public class AuthService {
     public AuthResponse.LoginResponse kakaoLogin(AuthRequest.KakaoLoginRequest request) {
 
         // 1. 회원 조회 후 없으면 생성
-        User user = userService.createKakaoUser(request);
+        User user = userCreateService.createKakaoUser(request);
 
         // 2. 관심 카테고리 매핑
-        userService.attachCategories(user, request.getCategoryIds());
+        userCreateService.attachCategories(user, request.getCategoryIds());
 
         // 3. Jwt 발급
         String accessToken = jwtTokenService.createAccessToken(user.getId());
@@ -59,10 +53,10 @@ public class AuthService {
     public AuthResponse.SignUpResponse signUp(AuthRequest.EmailSignUpRequest request) {
 
         // 1. 유저 생성
-        User user = userService.createUser(request);
+        User user = userCreateService.createUser(request);
 
         // 2. 관심 카테고리 매핑
-        userService.attachCategories(user, request.getCategoryIds());
+        userCreateService.attachCategories(user, request.getCategoryIds());
 
         return AuthConverter.toSignUpResponse(user);
     }
